+++ pow/src/main/java/tech/pegasys/teku/pow/Web3jEth1Provider.java
@@ -24,11 +24,15 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hyperledger.besu.plugin.services.MetricsSystem;
+import org.hyperledger.besu.plugin.services.metrics.Counter;
+import org.hyperledger.besu.plugin.services.metrics.LabelledMetric;
 import org.web3j.protocol.Web3j;
 import org.web3j.protocol.core.DefaultBlockParameter;
 import org.web3j.protocol.core.DefaultBlockParameterName;
 import org.web3j.protocol.core.Request;
 import org.web3j.protocol.core.Response;
+import org.web3j.protocol.core.Response.Error;
 import org.web3j.protocol.core.methods.request.EthFilter;
 import org.web3j.protocol.core.methods.request.Transaction;
 import org.web3j.protocol.core.methods.response.EthBlock;
@@ -37,6 +41,7 @@ import org.web3j.protocol.core.methods.response.EthChainId;
 import org.web3j.protocol.core.methods.response.EthLog;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.metrics.TekuMetricCategory;
 import tech.pegasys.teku.infrastructure.time.TimeProvider;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.pow.exception.RejectedRequestException;
@@ -50,16 +55,25 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
   private final String id;
   private final Web3j web3j;
   private final AsyncRunner asyncRunner;
+  private final LabelledMetric<Counter> requestCounter;
 
   public Web3jEth1Provider(
+      final MetricsSystem metricsSystem,
       final String id,
       final Web3j web3j,
       final AsyncRunner asyncRunner,
-      TimeProvider timeProvider) {
+      final TimeProvider timeProvider) {
     super(timeProvider);
     this.web3j = web3j;
     this.asyncRunner = asyncRunner;
     this.id = id;
+    requestCounter =
+        metricsSystem.createLabelledCounter(
+            TekuMetricCategory.BEACON,
+            "eth1_requests_total",
+            "Counter of the number of requests made to the eth1-endpoint, by endpoint ID and JSON-RPC method",
+            "endpoint",
+            "method");
   }
 
   @Override
@@ -124,11 +138,13 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
   @SuppressWarnings("rawtypes")
   private <S, T extends Response> SafeFuture<T> sendAsync(final Request<S, T> request) {
     try {
+      requestCounter.labels(id, request.getMethod()).inc();
       return SafeFuture.of(request.sendAsync())
           .thenApply(
               response -> {
                 if (response.hasError()) {
-                  throw new RuntimeException(response.getError().getMessage());
+                  final Error error = response.getError();
+                  throw new RejectedRequestException(error.getCode(), error.getMessage());
                 } else {
                   updateLastCall(Result.SUCCESS);
                   return response;
@@ -186,17 +202,7 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
   public SafeFuture<List<EthLog.LogResult<?>>> ethGetLogs(EthFilter ethFilter) {
     return sendAsync(web3j.ethGetLogs(ethFilter))
         .thenApply(EthLog::getLogs)
-        .thenApply(
-            logs -> {
-              if (logs == null) {
-                // We got a response from the node but it didn't include even an empty list
-                // of logs.  This happens with Infura when more than 10,000 log entries match
-                // so treat as an explicit rejection of the request to allow the requested block
-                // range to be reduced.
-                throw new RejectedRequestException("No logs returned by ETH1 node");
-              }
-              return (List<EthLog.LogResult<?>>) (List) logs;
-            });
+        .thenApply(logs -> (List<EthLog.LogResult<?>>) (List) logs);
   }
 
   @Override